Added simple file path correction assist
In Windows there is a context menu option to "Copy as path" which is a good feature. However, it adds a quotation mark at the beginning and end of the string. To overcome this, a simple file path correction assistance is added.

Also did minor changes in output formatting for prettier looks.