OutputDialog dynamic button display
OutputDialog now checks if the path supplied is to a file or to a directory. If if points to a file, all three buttons will be shown. If it points to a directory, only the Open Folder button will be shown because there is no point showing the Open File button when there is no file to open.