Fixed empty course info not caught in validation
When the course info is set to empty, the data pre-processor reads that its row contains an NA value hence it originally removes the row entirely, making it uncaught in the validation and could possibly break the system when its data is actually read. Now the data is simply turned into an empty string, allowing for validation to discover that there is a key with no value.