MS Word checking moved to software initialization
Checking if MS Word is installed is not performed when main is first called. The version string is then passed into the window constructor. This is to prevent opening and closing Word multiple times when the report generator is also opened multiple times. This is also to pre-load all needed checks, resulting in a much faster GUI experience post-loading.