[Feature] Moodle Database Accessor
* Add output text in PDF verifier returns

This is to support integration with GUI

* Added GUI for report integrity verifier

* Report generator UI improvement

- Added console logs where necessary
- Switched GUI to use a textbox instead. Hence, a mirror of the console log can be added into the textbox too.

* Validate grader report now uses GUI callback

* Added extra feedback log in GUI when generating

* MS Word checking moved to software initialization

Checking if MS Word is installed is not performed when main is first called. The version string is then passed into the window constructor. This is to prevent opening and closing Word multiple times when the report generator is also opened multiple times. This is also to pre-load all needed checks, resulting in a much faster GUI experience post-loading.

* Added Moodle database access module

* Added pymysql as a requirement

* Query console logging only on verbose mode

* Added pypdf as requirement

pypdf is used by metadata and integrity modules.

* Added a new database viewer GUI

The system uses webview so the data is output into an HTML file and opened by the webview engine.

* Added new GUI for Moodle database accessor