Bug Fixes and Improvements for v1.1.0
* Modularized HTML components

HTML components can now be placed anywhere within the program directory. This allows for customizability (such as themes) in the future.

* Updated method to test connection to Moodle DB

* Connection testing before initializing DB viewer

When the user clicks on the Moodle Database Viewer button, the launcher will do the checks without initializing the window first. This prevents the need to destroy an initialized window (unnecessarily initializing a window).

* Update Moodle database test connection output

The test connection test query output is now suppressible. By default it is suppressed to prevent unnecessary console clutter. Console printing is also now handled by the method instead of GUI.