Added menu bar on report generator window
With the growing list of toggles and buttons to access various features, a menu bar has been added to host a set of important buttons that are not necessarily used often. This design decision is to minimize clutter on the main window and to hide buttons that are more 'debug-y' in nature.